FileSystemStorage: make behavior on existing file configurable by optional property "onFileExists" fix #3983

+++ dcm4chee-arc-storage/src/main/java/org/dcm4chee/arc/storage/AbstractStorage.java
@@ -249,7 +249,7 @@ public abstract class AbstractStorage implements Storage {
 
     @Override
     public void revokeStorage(WriteContext ctx) throws IOException {
-        deleteObject(ctx.getStoragePath());
+        if (!ctx.isDeletionLock()) deleteObject(ctx.getStoragePath());
     }
 
     @Override
